rename dominant-cache-thrift-converter
